back/requirements.txt
Django~=4.0.3
djangorestframework~=3.13.1
djangorestframework-simplejwt